ai-agent-tutorial/requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
wikipedia
duckduckgo-search
